move accenting to only place where its used
