+++ kalite/i18n/tests/create_dummy_language_pack_tests.py
@@ -1,4 +1,4 @@
-import accenting
+from management import accenting
 import requests
 import zipfile
 from cStringIO import StringIO
